api/requirements.txt
flask
flask-cors
numpy
pillow
opencv-python
PyJWT
python-dotenv
